+++ etrm-system/etrm-backend/src/main/java/com/etrm/system/counterparty/CounterpartyRepository.java
@@ -2,6 +2,6 @@ package com.etrm.system.counterparty;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CounterpartyRepository extends JpaRepository<Counterparty, Long> {
+public interface CounterpartyRepository extends JpaRepository<Counterparty, Integer> {
     boolean existsByCpCodeIgnoreCase(String cpCode);
 }
