+++ etrm-system/etrm-backend/src/main/java/com/etrm/system/fieldpermission/FieldPermissionProfileRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface FieldPermissionProfileRepository extends JpaRepository<FieldPermissionProfile, Long> {
+public interface FieldPermissionProfileRepository extends JpaRepository<FieldPermissionProfile, Integer> {
     List<FieldPermissionProfile> findByScreenCodeAndIsActiveTrueOrderByProfileName(String screenCode);
     Optional<FieldPermissionProfile> findByProfileCode(String profileCode);
 }
