+++ etrm-system/etrm-backend/src/main/java/com/etrm/system/fieldpermission/ObjectLockRuleRepository.java
@@ -3,6 +3,6 @@ package com.etrm.system.fieldpermission;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ObjectLockRuleRepository extends JpaRepository<ObjectLockRule, Long> {
+public interface ObjectLockRuleRepository extends JpaRepository<ObjectLockRule, Integer> {
     List<ObjectLockRule> findByScreenCodeAndIsActiveTrueOrderBySortOrder(String screenCode);
 }
