+++ etrm-system/etrm-backend/src/main/java/com/etrm/system/fieldpermission/RoleFieldProfile.java
@@ -9,7 +9,7 @@ public class RoleFieldProfile {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "mapping_id")
-    private Long mappingId;
+    private Integer mappingId;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "role_id", nullable = false)
@@ -19,8 +19,8 @@ public class RoleFieldProfile {
     @JoinColumn(name = "profile_id", nullable = false)
     private FieldPermissionProfile profile;
 
-    public Long getMappingId() { return mappingId; }
-    public void setMappingId(Long mappingId) { this.mappingId = mappingId; }
+    public Integer getMappingId() { return mappingId; }
+    public void setMappingId(Integer mappingId) { this.mappingId = mappingId; }
     public UserRole getRole() { return role; }
     public void setRole(UserRole role) { this.role = role; }
     public FieldPermissionProfile getProfile() { return profile; }
