+++ etrm-system/etrm-backend/src/main/java/com/etrm/system/fieldpermission/RoleFieldProfileRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import java.util.List;
 
-public interface RoleFieldProfileRepository extends JpaRepository<RoleFieldProfile, Long> {
+public interface RoleFieldProfileRepository extends JpaRepository<RoleFieldProfile, Integer> {
 
     @Query("""
         SELECT rfp FROM RoleFieldProfile rfp
@@ -14,6 +14,6 @@ public interface RoleFieldProfileRepository extends JpaRepository<RoleFieldProfi
           AND rfp.profile.isActive = true
         """)
     List<RoleFieldProfile> findByRoleIdsAndScreenCode(
-            @Param("roleIds") List<Long> roleIds,
+            @Param("roleIds") List<Integer> roleIds,
             @Param("screenCode") String screenCode);
 }
