+++ etrm-system/etrm-backend/src/main/java/com/etrm/system/fieldpermission/ScreenFieldRegistryRepository.java
@@ -3,7 +3,7 @@ package com.etrm.system.fieldpermission;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ScreenFieldRegistryRepository extends JpaRepository<ScreenFieldRegistry, Long> {
+public interface ScreenFieldRegistryRepository extends JpaRepository<ScreenFieldRegistry, Integer> {
     List<ScreenFieldRegistry> findByScreenCodeAndIsActiveTrueOrderBySortOrder(String screenCode);
     List<String> findDistinctScreenCodeByIsActiveTrue();
 }
