+++ etrm-system/etrm-backend/src/main/java/com/etrm/system/legalentity/LegalEntityRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LegalEntityRepository extends JpaRepository<LegalEntity, Long> {
+public interface LegalEntityRepository extends JpaRepository<LegalEntity, Integer> {
     Optional<LegalEntity> findByEntityCodeIgnoreCase(String entityCode);
     boolean existsByEntityCodeIgnoreCase(String entityCode);
 }
